examples: Add time_decrypt.c for TPM 2 using IBMTSS 2
Add a script for taking time measurements of the decryption operation
of a TPM 2. It uses the TPM 2's NULL hierarchy to load the RSA key into
and then decrypts all messages using this key. It also counts the number
of decryption failures and returned messages of unexpected size and
reports them at the end.

Use the following command to compile:

 gcc time_decrypt.c -o time_decrypt -lcrypto -libmtss

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>